analytics/requirements.txt
apscheduler
flask==2.2.2
psycopg2-binary
flask_sqlalchemy
